% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{add_specifier}
\alias{add_specifier}
\title{Title}
\usage{
add_specifier(
  varnames,
  str = "i",
  n,
  sort = c("str1", "str2"),
  name_index = FALSE,
  unlist = FALSE
)
}
\arguments{
\item{varnames}{}

\item{str}{}

\item{n}{}

\item{sort}{}

\item{name_index}{}

\item{unlist}{}
}
\value{

}
\description{
Title
}
