% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{create_var_names}
\alias{create_var_names}
\title{Create variable names}
\usage{
create_var_names(var_name, str = "s", n, sep = "_", unlist = FALSE)
}
\arguments{
\item{var_name}{String, specifying the variable name.}

\item{str}{String, specifying the suffix to be added after the \code{var_name}.}

\item{n}{Numeric, specifying the number of variables.}

\item{sep}{String, specifying the seperatpr between \code{var_name} and \code{str}.}

\item{unlist}{Logical, specifying whether a "list" or "vector" should be returned.}
}
\value{

}
\description{
Create variable names
}
