% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{create_var_names2}
\alias{create_var_names2}
\title{Title}
\usage{
create_var_names2(
  str_name,
  str1 = "s",
  n1,
  str2 = "i",
  n2,
  sort = c("str1", "str2"),
  name_index = FALSE,
  unlist = FALSE
)
}
\arguments{
\item{str_name}{}

\item{str1}{}

\item{n1}{}

\item{str2}{}

\item{n2}{}

\item{sort}{}

\item{name_index}{}

\item{unlist}{}
}
\value{

}
\description{
Title
}
