% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{add_specifier}
\alias{add_specifier}
\title{Add further specifications to variable names}
\usage{
add_specifier(
  var_names,
  str = "i",
  n,
  sort = c("previous", "current"),
  name_index = FALSE,
  unlist = FALSE
)
}
\arguments{
\item{var_names}{List or vector of variable names}

\item{str}{}

\item{n}{}

\item{sort}{}

\item{name_index}{}

\item{unlist}{}
}
\value{

}
\description{
Add further specifications to variable names
}
