% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-var-names.R
\name{create_var_names}
\alias{create_var_names}
\title{Title}
\usage{
create_var_names(str_name, str = "s", n)
}
\arguments{
\item{str_name}{}

\item{str}{}

\item{n}{}
}
\value{

}
\description{
Title
}
